packages/hydrofetch-dashboard-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hydrofetch-dashboard-api"
version = "0.1.0"
description = "FastAPI backend for the Hydrofetch local monitoring dashboard"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0",
    "pyarrow>=23.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hydrofetch_dashboard_api"]
